+++ editor/src/main/kotlin/de/fatox/meta/EditorMeta.kt
@@ -11,7 +11,7 @@ import de.fatox.meta.injection.MetaInject.Companion.lazyInject
 import de.fatox.meta.modules.MetaEditorModule
 import de.fatox.meta.modules.MetaUIModule
 import de.fatox.meta.screens.MetaEditorScreen
-import de.fatox.meta.screens.SplashScreen
+import de.fatox.meta.api.SplashScreen
 
 class EditorMeta(posM: PosModifier) : Meta(posM) {
